refactor: migrate sts-generator to itty router
