+++ packages/workers/ens-resolver/package.json
@@ -13,7 +13,7 @@
     "worker:deploy": "wrangler deploy"
   },
   "dependencies": {
-    "itty-router": "4.0.9",
+    "itty-router": "^4.0.9",
     "viem": "^0.3.19"
   },
   "devDependencies": {
