+++ packages/workers/oembed/.eslintrc.js
@@ -3,6 +3,5 @@ module.exports = {
   extends: ['node'],
   rules: {
     'import/no-anonymous-default-export': 'off'
-  },
-  ignorePatterns: ['generated.ts']
+  }
 };
