+++ packages/workers/snapshot-relay/package.json
@@ -19,7 +19,7 @@
     "@apollo/client": "^3.7.15",
     "@lenster/lib": "workspace:*",
     "@tsndr/cloudflare-worker-jwt": "^2.2.1",
-    "itty-router": "4.0.9",
+    "itty-router": "^4.0.9",
     "viem": "^0.3.19"
   },
   "devDependencies": {
