+++ packages/workers/sts-generator/package.json
@@ -13,7 +13,8 @@
     "worker:deploy": "wrangler deploy"
   },
   "dependencies": {
-    "@aws-sdk/client-sts": "^3.342.0"
+    "@aws-sdk/client-sts": "^3.342.0",
+    "itty-router": "^4.0.9"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20230518.0",
