+++ packages/workers/sts-generator/src/index.ts
@@ -1,75 +1,35 @@
-import { AssumeRoleCommand, STSClient } from '@aws-sdk/client-sts';
+import { createCors, error, json, Router } from 'itty-router';
 
-interface EnvType {
-  EVER_ACCESS_KEY: string;
-  EVER_ACCESS_SECRET: string;
-}
+import getStsToken from './handlers/getStsToken';
+import type { Env } from './types';
 
-const headers = {
-  'Access-Control-Allow-Origin': '*',
-  'Content-Type': 'application/json'
-};
-const bucketName = 'lenster-media';
-const everEndpoint = 'https://endpoint.4everland.co';
+const { preflight, corsify } = createCors({
+  origins: ['*'],
+  methods: ['HEAD', 'GET']
+});
 
-const params = {
-  DurationSeconds: 900,
-  Policy: `{
-    "Version": "2012-10-17",
-    "Statement": [
-      {
-        "Effect": "Allow",
-        "Action": [
-          "s3:PutObject",
-          "s3:GetObject"
-        ],
-        "Resource": [
-          "arn:aws:s3:::${bucketName}/*"
-        ]
-      }
-    ]
-  }`
-};
+const router = Router();
 
-async function handleRequest(request: Request, env: EnvType) {
+router.all('*', preflight);
+router.get('/', getStsToken);
+
+const routerHandleStack = (request: Request, env: Env, ctx: ExecutionContext) =>
+  router.handle(request, env, ctx).then(json);
+
+const handleFetch = async (
+  request: Request,
+  env: Env,
+  ctx: ExecutionContext
+) => {
   try {
-    const accessKeyId = env.EVER_ACCESS_KEY;
-    const secretAccessKey = env.EVER_ACCESS_SECRET;
-
-    const stsClient = new STSClient({
-      endpoint: everEndpoint,
-      region: 'us-west-2',
-      credentials: { accessKeyId, secretAccessKey }
-    });
-
-    const data = await stsClient.send(
-      new AssumeRoleCommand({
-        ...params,
-        RoleArn: undefined,
-        RoleSessionName: undefined
-      })
-    );
-
-    return new Response(
-      JSON.stringify({
-        success: true,
-        accessKeyId: data.Credentials?.AccessKeyId,
-        secretAccessKey: data.Credentials?.SecretAccessKey,
-        sessionToken: data.Credentials?.SessionToken
-      }),
-      { headers }
-    );
-  } catch (error) {
-    console.error('Failed to generate STS token', error);
-    return new Response(
-      JSON.stringify({ success: false, message: 'Something went wrong!' }),
-      { headers }
-    );
+    return await routerHandleStack(request, env, ctx);
+  } catch (error_) {
+    console.error('Failed to handle request', error_);
+    return error(500);
   }
-}
+};
 
 export default {
-  async fetch(request: Request, env: EnvType) {
-    return await handleRequest(request, env);
-  }
+  fetch: (request: Request, env: Env, context: ExecutionContext) =>
+    handleFetch(request, env, context).then(corsify)
 };
