+++ packages/workers/sts-generator/src/types.ts
@@ -1,0 +1,4 @@
+export interface Env {
+  EVER_ACCESS_KEY: string;
+  EVER_ACCESS_SECRET: string;
+}
